+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.5.1] - 2026-07-23
 
 ### Changed
 - JWT-VC / JWT-VP signing: JOSE protected header now includes `typ` `vc+jwt` / `vp+jwt` per [VC-JOSE-COSE](https://www.w3.org/TR/vc-jose-cose/)
+
 
 ## [2.5.0] - 2026-07-03